waypoint generation comparisons
This file allows us to compare the waypoint generation of a few different networks. I still need to check a few things. Minor changes from our current waypoint generation code. It is still running as I upload it, so I still need to test it out, but I wanted to put it here for now.